third_party/blink/web_tests/fast/files/xhr-response-blob-expected.txt

Test that XHR.responseType = 'blob' gives you back a Blob.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS xhr.responseType is "blob"
PASS successfullyParsed is true

TEST COMPLETE

PASS xhr.response is null
PASS xhr.response is null
PASS xhr.response instanceof Blob is true
PASS xhr.response.type is "text/plain"
PASS xhr.responseType is "blob"
FAIL xhr.response instanceof Blob should be true. Was false.
FAIL xhr.response.type should be . Threw exception TypeError: Cannot read properties of null (reading 'type')
PASS xhr.responseType is "blob"
PASS xhr.response is null
PASS xhr.response instanceof Blob is true
PASS xhr.response.type is "text/plain"
